update GVapp to correspond to updates in GV
