+++ bin/genevalidatorapp
@@ -55,9 +55,6 @@ BNR
   o.string '-s', '--set',
            'Set configuration value in the config file'
 
-  o.string '-l', '--list_dbs',
-           'List BLAST databases'
-
   o.string '-b', '--blast_bin',
            'Load BLAST+ binaries from this directory'
 
@@ -84,7 +81,7 @@ if opts.help?
 end
 
 if opts.version?
-  require 'GeneValidatorApp/version'
+  require 'genevalidatorapp/version'
   puts GeneValidatorApp::VERSION
   exit
 end
@@ -105,7 +102,7 @@ clean_opts = lambda do |hash|
   hash
 end
 
-require 'GeneValidatorApp'
+require 'genevalidatorapp'
 begin
   GeneValidatorApp.init clean_opts[opts.to_hash]
 rescue SystemExit => e
@@ -114,8 +111,6 @@ rescue SystemExit => e
   puts '    See https://github.com/IsmailM/GeneValidatorApp for more help'
   exit e.status
 end
-
-puts GeneValidatorApp::Database.all? if opts.list_databases?
 
 GeneValidatorApp.send(:write_config_file) if opts.set?
