+++ genevalidatorapp.gemspec
@@ -1,7 +1,7 @@
 # coding: utf-8
 lib = File.expand_path('../lib', __FILE__)
 $LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
-require 'GeneValidatorApp/version'
+require 'genevalidatorapp/version'
 
 Gem::Specification.new do |spec|
   spec.name          = 'genevalidatorapp'
@@ -10,7 +10,7 @@ Gem::Specification.new do |spec|
   spec.email         = ['Ismail.Moghul@gmail.com']
   spec.summary       = 'A Web App wrapper for GeneValidator.'
   spec.description   = 'A Web App wrapper for GeneValidator, a program for validating gene predictions.'
-  spec.homepage      = 'https://github.com/wurmlab/GeneValidatorApp'
+  spec.homepage      = 'https://github.com/wurmlab/genevalidatorapp'
   spec.license       = 'AGPL'
 
   spec.files         = `git ls-files -z`.split("\x0")
@@ -25,8 +25,8 @@ Gem::Specification.new do |spec|
   spec.add_development_dependency 'rspec', '~> 3.2'
   spec.add_development_dependency 'rack-test', '~> 0.6'
 
+  spec.add_dependency 'genevalidator', '~>1.7'
   spec.add_dependency 'bio', '~>1.4'
-  spec.add_dependency 'genevalidator', '~>1.5'
   spec.add_dependency 'sinatra', '~>1.4'
   spec.add_dependency 'sinatra-contrib', '~>1.4'
   spec.add_dependency 'sinatra-cross_origin', '~> 0.3'
@@ -43,7 +43,7 @@ Gem::Specification.new do |spec|
 
     $ genevalidatorapp [options]
 
-  Visit https://github.com/wurmlab/GeneValidatorApp for more information.
+  Visit https://github.com/wurmlab/genevalidatorapp for more information.
 ------------------------------------------------------------------------
 
 INFO
